requirements.txt
click==8.0.4
colorama==0.4.4
gensim==3.8.3
joblib==1.1.1
nltk==3.7
numpy==1.22.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2021.3
regex==2022.1.18
scikit-learn==1.2.0
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
threadpoolctl==3.1.0
tqdm==4.62.3
requests==2.27.1
networkx==2.6.3
seaborn
datasets
hydra-core
docker
virtualenv
grakel
